requirements.txt
openai>=1.50.0
openai-agents>=0.2.0
mcp[cli]>=1.0.0
python-dotenv>=1.0.1
requests>=2.32.3
agents
mcp